优化功能，不强制依赖es

+++ roncoo-education-course/roncoo-education-course-service/src/main/java/com/roncoo/education/course/job/CourseJob.java
@@ -1,6 +1,7 @@
 package com.roncoo.education.course.job;
 
 import cn.hutool.core.collection.CollUtil;
+import cn.hutool.core.util.ObjectUtil;
 import com.roncoo.education.common.core.enums.PutawayEnum;
 import com.roncoo.education.common.core.enums.StatusIdEnum;
 import com.roncoo.education.common.core.tools.BeanUtil;
@@ -30,28 +31,30 @@ import java.util.List;
 @Component
 public class CourseJob {
 
+    @Autowired(required = false)
+    private ElasticsearchRestTemplate elasticsearchRestTemplate;
     @Autowired
     private CourseDao courseDao;
-    @Autowired
-    private ElasticsearchRestTemplate elasticsearchRestTemplate;
 
     /**
      * 每天凌晨5点执行一次
      */
     @XxlJob("courseJobHandler")
     public void course() {
-        CourseExample example = new CourseExample();
-        example.createCriteria().andStatusIdEqualTo(StatusIdEnum.YES.getCode()).andIsPutawayEqualTo(PutawayEnum.UP.getCode());
-        List<Course> courseList = courseDao.listByExample(example);
-        if (CollUtil.isNotEmpty(courseList)) {
-            List<IndexQuery> queries = new ArrayList<>();
-            for (Course course : courseList) {
-                EsCourse esCourse = BeanUtil.copyProperties(course, EsCourse.class);
-                IndexQuery query = new IndexQueryBuilder().withObject(esCourse).build();
-                queries.add(query);
+        if (ObjectUtil.isNotNull(elasticsearchRestTemplate)) {
+            CourseExample example = new CourseExample();
+            example.createCriteria().andStatusIdEqualTo(StatusIdEnum.YES.getCode()).andIsPutawayEqualTo(PutawayEnum.UP.getCode());
+            List<Course> courseList = courseDao.listByExample(example);
+            if (CollUtil.isNotEmpty(courseList)) {
+                List<IndexQuery> queries = new ArrayList<>();
+                for (Course course : courseList) {
+                    EsCourse esCourse = BeanUtil.copyProperties(course, EsCourse.class);
+                    IndexQuery query = new IndexQueryBuilder().withObject(esCourse).build();
+                    queries.add(query);
+                }
+                elasticsearchRestTemplate.indexOps(EsCourse.class).delete();
+                elasticsearchRestTemplate.bulkIndex(queries, IndexCoordinates.of(EsCourse.COURSE));
             }
-            elasticsearchRestTemplate.indexOps(EsCourse.class).delete();
-            elasticsearchRestTemplate.bulkIndex(queries, IndexCoordinates.of(EsCourse.COURSE));
         }
         XxlJobHelper.handleSuccess("完成");
     }
